small include fix added mass make 8bit file

+++ make8bit.py
@@ -5,7 +5,7 @@ import sys
 if __name__ == '__main__':
 
     # Read file
-    inputimg = np.array(get_itk_array(sys.argv[1]))
+    inputimg = np.array(get_itk_array(sys.argv[1]), dtype='int16')
 
     inputimg = (inputimg/255).astype('uint8')
